Use int64 for power calculations, not uint16
Fixes an overflow when computing the maximum amount of support a
proposal could have had. There's no reason to use uint16 here anyways,
even if power values are guaranteed to fit.

+++ gpbft/message_builder.go
@@ -19,7 +19,7 @@ type MessageBuilder struct {
 }
 
 type powerTableAccessor interface {
-	Get(ActorID) (uint16, PubKey)
+	Get(ActorID) (int64, PubKey)
 }
 
 type SignerWithMarshaler interface {
